Update required nrfconnect engine to ^2.3.0
The PPK depends on RTT, which was introduced in pc-nrfjprog-js 1.2.0.
The pc-nrfconnect-core versions prior to 2.3.0 came with an older
version of pc-nrfjprog-js, so the app will not work for core versions
below 2.3.0.

Updating the engine config so that users get a warning when trying to
launch the app on older pc-nrfconnect-core versions.