filtering: best suitable filter settings as of yet
These settings has been tested with all DK's, both dongles and the
desktop2 keyboard. We need to include quite a lot of spikes to
get the most accurate measurement, but filter some to not get
too many spikes. This was a good middle point where accuracy
for all boards are very good, but still contains some false, higher
spikes in the measurement set.