Restructure display logic
While the logic whether to show the display options was in that
component, the same logic for the save button below is in the sidepanel.
Now both are combined in the sidepanel. And the “when a device is open
show one thing, otherwise another thing” logic is better expressed with
a ternary than with `deviceOpen ||` and afterwards `deviceOpen &&`.