Add new handle to offset trigger window for ppk2
Previous pre/post-sampling solution restricted users to a hardcoded
value. Replace this functionality with a new handler that allows user to
shift the trigger window to display the area of interest. This will only
work on PPK2, because of PPK1 hardware limitations.

We also add a button to hide the chart grid to reduce visual clutter if
the user should want.