Rework digital channel data accumulation
This changes the algorithm to make it more performant.

When we have more data than the current resolution can display
(step > 1) we combine data into a single data point.

For digital channels the old algorithm looked whether within the step we
had a change (from 0 to 1 or vice versa) and only used that as a data
point. While this minimised the data points (which is helpful for
chart.js), this computation was expensive.

The new algorithm is more similar to the algorithm also used for the
analog data: We look whether within the step the channel was ever 0, 1
or both. If so we set a data point low, high or both.

While there are corner cases where this means a line going up or down
can be a pixel wider, it does not really change the shape of the graph.

Another change in the algorithm was that previously we went through the
data within the step multiple times: Once for the analog data and once
for every digital channel. This is now changed to only go through it
once and look at all channels in this one go.

Rendering the Chart is the most costly portion of tree reconciliation,
e.g. when new data comes in or users scroll.

With a fixed data set rendering the Chart component usually took around
30ms, of which combining the data into a single data point was the
biggest chunk and took around 20ms. With this new algorithm rendering
the data is usually down below 18ms, of which combining the data took
below 8ms.

The new algorithm also leads to more data points to be displayed but in
practice the performance penality for that was lower than the speed up
because of the different algorithm. Rendering a digital channel with
more or less data was always below 0.2ms.