Switch bit accumulator algoritm
Instead of storing whether the bit was at least once on/off, count how
often it was on. At the end we can decide based on that whether it was
Always off (counter = 0), always on (counter = max) or sometimes each
(counter in between). We also have to check whether we really have data
for the current timestamp, which we can detect by looking at the analog
data.

This speeds up the calculation by about 25%.