Enhance performance by throttling wheel events
There can be an overwhelming amount of wheel events which hurts the
performance. This commit enhances that situation by utilising
`requestAnimationFrame`.

If multiple events occured before the next animation frame is drawn,
the computations is only done once, thus the system is not overwhelmed
so easily by the events.

And the computations are done for the last event, not for older ones,
which should also make the system appear more responsive.

The main part of the algorithm is not changed and just extracted from
`zoomPan.wheelHandler` to `processWheelEvents`.