Display averaged bit data as an area
When in an averaged timespan a bit was both 0 and 1 we previously showed
the line neither up nor down but in the middle. With this change we show
an area instead, filling the whole room from up to down, thus
symbolising that it is unknown at which moments in this timespan the
signal was 0 or 1.