Refactor load and save handlers.
Instead of returning an object that has to be initialised and which
exposes functionality that should not be exposed we instead create a
function and pass the necessary data through.
We also add the triggerstate to the savefile.
Lastly, this commit improves error handling such that the user is
notified by an error in the log window when a file cannot be saved or
loaded.