Invert bit sequence when exporting data.
Currently the bit sequence is inverted, i.e. the bit representing
channel D0 is in the last position and D7 is in the first position. So
we need to fix this to display it properly in the exported CSV files.

(cherry picked from commit 506d7e15fe4ef25467855cba9ca6c12599412d74)