Update to shared v4.25.0
That new version of shared also updated the version of css-loader from
3.5.3 to 5.2.6, which changed how to import variabled from CSS into JS.

If one wants to do this, the (S)CSS file needs to be explicitly compiled
with ICSS (Interoperable CSS) support. With the default configuration
this is achieved by naming the CSS files that export variables to JS
with a `.icss.` in the filename.