Add radio button for area to export
We have received some feedback that customers expected the full sample
array to be exported by default. When it did not it's easy to think the
app is broken. The radio buttons make it clear what option is selected,
and provides the customer with the ability to change what to export.