Do not reset index when sampling is complete
Several new features depend on the options.index pointing at the
accurate place in the buffer. I see no reason for setting the index to
zero since we do not have a circular buffer.

At least note that the functions in globals.js; `timestampToIndex` and
`indexToTimestamp` rely on timestamp and index being somehow synchronized.