Remove device filter in order to display PPK2
PPK2 did not show up in device filter when the firmware was not exactly
the correct one. The firmware may vary, for example if the device is
just from factory, or if customer has flashed a custom fw.

We want to allow customers to be able to select any device, and then
we'll flash the correct firmware when the device is selected.