Add windowDuration on real time trigger
To properly update the window width, in order to display the correct
amount of samples, we also need to update chart state with
windowDuration.

To further improve this, we could update the chart with the set length
immediately after adjusting the length in the side panel. As of now, it
only updates on a trigger event.

Known Issue: PPK1 does update windowDuration appropriately, this is not
a regression for upcoming release.