Reduce maxVoltageCap when PPK1 is selected
Because of the max voltage cap feature that was implemented earlier, the
PPK1 is also able to select a higher VDD then it is suppose to. This
commit fix that again, with a minor drawback.

Since the redux store need to be updated, during a session the store
will be updated to use the PPK1 upper limit. Meaning that if a PPK2 is
selected after PPK1 on the same session, the PPK2 will also be capped.