change: restrict zoom
When max zoom out, or max zoom in has been reached, the window would
pane in the direction of the mouse, relative to the middle of the
screen. In theory, if the mouse pointer would be perfectly in the middle,
nothing would happen.

This commit restrict the pane movement on zoom actions when the zoom
in or zoom out limit is reached.