Draw minimap
The minimap is very blurry, and not particularily optimized. Could still
work out fine as long as we do not update the map very often.

Additionally, I think we can store the coordinate and index of the last
updated sample, and then iterate from there updating the minimap. This
is why I also think it's good to have the size of the minimap fixed.