checkpoint: updating, paning with edges.
Paning works quite well, and you cannot pane further than to the edge of
the samples. Meaning it stops on 0 paning left and at the options.index
at the right.

Updating during sampling works fine, but should be reviewed. Now it will
simply updating per ~100_000th sample, meaning it will add about 100
data points on each update.

The resolution of the minimap is bad, since it only looks at every
1000th sample, and can potentially skip important sample data, which
will then not be visible in the minimap.