Improve minimap by showing min/max of all points
In order to better capture fall and rise of the measurements, we iterate
over all the samples, and inside an interval we capture the minimum and
maximum value. This way we draw a vertical line on the given x that
represent the whole interval of samples.

This commit also gives the minimap a better layout in terms of padding
it to fit the rest of the content. In order to show the slider
correctly, we have to calculate the offset made by the padding.