Ensure that ZoomPan only run inside allowable window
Zoom function should always contain either part of the sample, or it
should contain the entire span of the sample, and only then can it show
more than the sample (on the right side).

Similarly, when the window displays more than the entire span of the
sample, and the user zooms in, then the window should continue to span
the entire sample until the window contains 100% of the sample.

The pan function should only pan when the window is within the span of
the sample. If the window shows more than the entire span of the sample,
then it should do nothing.