Move y-axis settings into a dialog
In order to adapt the ChartTop to a more responsive design, I've decided
to put some of the settings into a dialog. This avoids overflowing the
top of the chart.

However, it is still possible to make the screen so small that some of
the components in the top will overlap. A solution is to set a
min-width, but this would instead hide certain options.