fix: The chart repeated itself when zooming out
To reproduce, you would sample for just a few seconds, on for
example 1000 samples per second, and when you zoomed out, it would show
the same chart repeated across the window.

The logic that seemed to produce this bug is a remains from when the app
used a circular buffer in order to store the samples.