Default to NaN if value is undefined
The perceived performance was clearly affected by getting several
undefined values. Hence, give the variable v the NaN value when n is out
of bounds of data[].

The performance may be affected by how many `bitDataProcessor.processBits(n)`
is called, which will only improve the performance in the beginning,
because that is when you'd encountered getting undefined values.

The reason why you would get undefined values at all, is because of a
limited size of the options.data (the sample buffer), which means that
you would have to scale down the total sampling time significantly. This
in combination of zooming out more than what the total sampling time is,
would lead the dataAccumulator to attempt to retrieve values that are
outside of the length of options.data.