fix: Minimap was not properly removed
Make sure that the minimap is properly hidden/removed when you do
multiple samples.

The bug was present because we only checked if the
options.data had entries, but we need to make sure a minimap is actually
generated before displaying the minimap.

This commit also makes sure that the Minimap is only rendered if the App
is rendering the Data Logger pane, as this functionality should not be
present in the Real Time pane.