fix: SELECTION statistics with negative time delta
On using the right handle to adjust the selection, you could drag
it passed the left handle, making the time negative. This commit
makes sure that `begin < end`, so that the delta is always `> 0`.

Additionally, you could extend the left handle (=begin) passed 0
on the X axis, so that you could make a SELECTION that would be
larger than the data set. Hence begin should always be `>= 0`.